Add built-in WHPX VMM; harden input handling across NFS/VMM/shell-exec paths
- New whpx_vmm.c/h: minimal Windows Hypervisor Platform VMM as a QEMU-free
  path for persistent Alpine sessions (virtio-blk, serial console), wired
  into qemu_linux.c's run_persist_session_whpx() with Job Object cleanup
  for orphaned QEMU/nfs-worker processes.
- whpx_vmm.c: bound every guest-physical-address dereference in the
  virtio-blk path against actual RAM size, and cap/validate descriptor
  chain walks -- closes an OOB host-memory read/write and an infinite-loop
  DoS reachable from a corrupted or malicious virtio ring.
- vdisknfsd.c/nfsclient.c: fix 32-bit XDR length-padding overflow that let
  a crafted RPC length defeat the bounds check and over-read past the
  received buffer; same fix applied to the WRITE handler's stream-position
  arithmetic.
- vdisknfsd.c: stop silently aliasing new files to the export root once
  the id table fills up; now returns NFS3ERR_NOSPC and cleans up the
  just-created file/dir instead of handing back a colliding handle.
- fs_memfs.c: reject an out-of-range/overflowing write offset before it
  can bypass the grow/quota check and write past the backing allocation.
- bridge.c, qemu_linux.c, main.c: a crafted SMB password, apk package
  name, or CLI argument can no longer break out of its intended quoting
  (guest shell command lines, and the elevated re-invocation's argv).

+++ nfsclient.c
@@ -80,19 +80,22 @@ static unsigned long long rd_u64(xreader_t *r) {
 static unsigned int rd_bytes(xreader_t *r, unsigned char *out, unsigned int outcap) {
     unsigned int n = rd_u32(r);
     if (r->error) return 0;
-    unsigned int padded = (n + 3) & ~3u;
-    if (r->pos + padded > r->len) { r->error = 1; return 0; }
+    // 64-bit: n comes from the peer and near UINT_MAX wraps (n+3) back to
+    // ~0 in 32-bit arithmetic, defeating the bounds check and letting
+    // memcpy read up to outcap bytes past the real received buffer.
+    unsigned long long padded = ((unsigned long long)n + 3) & ~3ull;
+    if ((unsigned long long)r->pos + padded > (unsigned long long)r->len) { r->error = 1; return 0; }
     unsigned int copy = n < outcap ? n : outcap;
     memcpy(out, r->data + r->pos, copy);
-    r->pos += padded;
+    r->pos += (size_t)padded;
     return copy;
 }
 static void rd_skip_bytes(xreader_t *r) {
     unsigned int n = rd_u32(r);
     if (r->error) return;
-    unsigned int padded = (n + 3) & ~3u;
-    if (r->pos + padded > r->len) { r->error = 1; return; }
-    r->pos += padded;
+    unsigned long long padded = ((unsigned long long)n + 3) & ~3ull;
+    if ((unsigned long long)r->pos + padded > (unsigned long long)r->len) { r->error = 1; return; }
+    r->pos += (size_t)padded;
 }
 static void rd_fh(xreader_t *r, nfs_fh_t *fh) {
     fh->len = rd_bytes(r, fh->data, sizeof(fh->data));
